[IMP] Adjust readme, add vacuum cron
