+++ user_activity_log/README.rst
@@ -17,16 +17,34 @@ User Activity Log
     :target: https://github.com/qrtl/nrq-custom/tree/12.0/user_activity_log
     :alt: qrtl/nrq-custom
 
-|badge1| |badge2| |badge3|
+|badge1| |badge2| |badge3| 
 
 This module does the following:
 
 - Allows to track specific user's operations.
 
+This module was developed taking the ideas from user_recent_log module by Aktive Software.
+
 **Table of contents**
 
 .. contents::
    :local:
+
+Installation
+============
+
+Besides installing this module, you need to add it to your addons, and load it as a server-wide module.
+
+This can be done with the server_wide_modules parameter in /etc/odoo.conf or with the --load command-line parameter
+Please put this module before the web module as core api method will be monkey patched.
+
+server_wide_modules = "use_activity_log,web"
+
+Usage
+=====
+
+For user you wish to track their activities, open the user form and click the "Track" button.
+A new menu "User Activity Log(s)" is added under the Settings > Technical > Security.
 
 Bug Tracker
 ===========
@@ -45,6 +63,7 @@ Authors
 ~~~~~~~
 
 * Quartile Limited
+* Aktiv Software
 
 Maintainers
 ~~~~~~~~~~~
