+++ user_activity_log/models/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from . import api
 from . import res_users
 from . import user_activity_log
