+++ user_activity_log/readme/DESCRIPTION.rst
@@ -1,3 +1,5 @@
 This module does the following:
 
 - Allows to track specific user's operations.
+
+This module was developed taking the ideas from user_recent_log module by Aktive Software.
