+++ user_activity_log/readme/USAGE.rst
@@ -1,0 +1,2 @@
+For user you wish to track their activities, open the user form and click the "Track" button.
+A new menu "User Activity Log(s)" is added under the Settings > Technical > Security.
